Travis-CI integration (#61)
Travis-CI build created and configured to download and install dependencies, and build for osx and linux. It runs tests, generates documentation, pushes documentation to tmxparser-docs, and pushes build artifacts to tmxparser-build. The two main branches in tmxparser-build are linux and osx for tar.gz's containing libraries and include for each. Builds will be triggered upon master pushes.